requirements.txt
pandas
numpy
google-play-scraper
psycopg2-binary
python-dotenv
matplotlib
seaborn
wordcloud
nltk
textblob
jupyter